+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @skild/core
 
+## 0.2.0
+
+### Minor Changes
+
+- cli with stats tracking
+
 ## 0.1.4
 
 ### Patch Changes